Moved common functions to lib dir.
